@@ -4,13 +4,12 @@ Auto-generated by: https://github.com/pmndrs/gltfjsx
 
 import React, { useRef } from "react";
 import { useGLTF } from "@react-three/drei";
+import wFlange from "../../assets/gltf/structure/w-flange.gltf";
 
 export default function Model({ position = [0, 0, 0], ...props }) {
   const group = useRef();
 
-  const { nodes, materials } = useGLTF(
-    "../../assets/gltf/structure/w-flange.gltf"
-  );
+  const { nodes, materials } = useGLTF(wFlange);
   return (
     <group ref={group} {...props} dispose={null}>
       <mesh
@@ -19,6 +18,7 @@ export default function Model({ position = [0, 0, 0], ...props }) {
       />
     </group>
   );
+  return <group />;
 }
 
-useGLTF.preload("../../assets/gltf/structure/w-flange.gltf");
+useGLTF.preload(wFlange);
